# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.BookResponse:
    properties:
      author:
        example: Tere Liye
        type: string
      book_id:
        example: b2a0f3c4-5d8e-4c1b-9f7e-2d3f4e5a6b7c
        type: string
      isbn:
        example: 978-3-16-148410-0
        type: string
      published_at:
        example: "2016-01-28"
        type: string
      publisher:
        example: Gramedia
        type: string
      stock:
        example: 200
        type: integer
      title:
        example: Hujan
        type: string
    type: object
  model.CreateBookRequest:
    properties:
      author:
        example: Tere Liye
        type: string
      isbn:
        example: "9783161484100"
        type: string
      published_at:
        example: "2016-01-28"
        type: string
      publisher:
        example: Gramedia
        type: string
      stock:
        example: 200
        minimum: 0
        type: integer
      title:
        example: Hujan
        type: string
    required:
    - author
    - isbn
    - published_at
    - publisher
    - stock
    - title
    type: object
  model.PaginatedResponse-model_BookResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.BookResponse'
        type: array
      links:
        $ref: '#/definitions/model.PaginationLinks'
      meta:
        $ref: '#/definitions/model.PaginationMeta'
    type: object
  model.PaginationLinks:
    properties:
      first:
        example: /api/v1/books?offset=0&limit=10
        type: string
      last:
        example: /api/v1/books?offset=90&limit=10
        type: string
      next:
        example: /api/v1/books?offset=10&limit=10
        type: string
      prev:
        example: /api/v1/books?offset=0&limit=10
        type: string
      self:
        example: /api/v1/books?offset=0&limit=10
        type: string
    type: object
  model.PaginationMeta:
    properties:
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
  model.UpdateBookRequest:
    properties:
      author:
        example: Tere Liye
        type: string
      book_id:
        type: string
      isbn:
        example: "9783161484100"
        type: string
      published_at:
        example: "2016-01-28"
        type: string
      publisher:
        example: Gramedia
        type: string
      stock:
        example: 200
        minimum: 0
        type: integer
      title:
        example: Hujan
        type: string
    required:
    - book_id
    type: object
info:
  contact: {}
  description: Backend API service for Book Stock Manager
  title: Book Stock Manager API
  version: 0.0.1
paths:
  /books:
    get:
      consumes:
      - application/json
      description: Get a list of books with pagination
      parameters:
      - description: 'Page offset (default: 0)'
        in: query
        name: offset
        type: integer
      - description: 'Page limit (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Books information with pagination metadata
          schema:
            $ref: '#/definitions/model.PaginatedResponse-model_BookResponse'
        "400":
          description: Invalid query parameters
          schema:
            type: string
        "500":
          description: Failed to get books
          schema:
            type: string
      summary: Get books with pagination
      tags:
      - books
    patch:
      consumes:
      - application/json
      description: Update book information partially
      parameters:
      - description: Request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdateBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated book
          schema:
            $ref: '#/definitions/model.BookResponse'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Book not found
          schema:
            type: string
        "500":
          description: Failed to update book
          schema:
            type: string
      summary: Update book
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Create a new book with the provided information
      parameters:
      - description: Request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreateBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Book created
          schema:
            $ref: '#/definitions/model.BookResponse'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Failed to create book
          schema:
            type: string
      summary: Create a new book
      tags:
      - books
  /books/{book_id}:
    get:
      consumes:
      - application/json
      description: Get a book information by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book information
          schema:
            $ref: '#/definitions/model.BookResponse'
        "400":
          description: Book ID is required
          schema:
            type: string
        "404":
          description: Book not found
          schema:
            type: string
        "500":
          description: Failed to get book by ID
          schema:
            type: string
      summary: Get book by ID
      tags:
      - books
  /books/{id}:
    delete:
      consumes:
      - application/json
      description: Delete book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Book ID is required
          schema:
            type: string
        "404":
          description: Book not found
          schema:
            type: string
        "500":
          description: Failed to delete book
          schema:
            type: string
      summary: Delete book
      tags:
      - books
  /books/isbn/{isbn}:
    get:
      consumes:
      - application/json
      description: Get a book information by ISBN
      parameters:
      - description: ISBN
        in: path
        name: isbn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book information
          schema:
            $ref: '#/definitions/model.BookResponse'
        "400":
          description: ISBN is required
          schema:
            type: string
        "404":
          description: Book not found
          schema:
            type: string
        "500":
          description: Failed to get book by ISBN
          schema:
            type: string
      summary: Get book by ISBN
      tags:
      - books
swagger: "2.0"